apps/backend/pyproject.toml
[tool.poetry]
name = "kcardswap-backend"
version = "0.1.0"
description = "KCardSwap 後端 API 服務"
authors = ["KCardSwap Team <team@kcardswap.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.1"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.2"  # pin <4.0.0 to avoid passlib bcrypt backend __about__ incompatibility
python-multipart = "^0.0.18"
sqlalchemy = "^2.0.45"
google-auth = ">=2.15.0,<2.42.0"
google-auth-oauthlib = "^1.2.3"
google-auth-httplib2 = "^0.2.1"
asyncpg = "^0.31.0"
alembic = "^1.17.2"
httpx = "^0.26.0"
google-cloud-storage = "^3.7.0"
firebase-admin = "^6.5.0"
injector = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --disable-warnings"
markers = [
    "gcs_smoke: GCS smoke tests that require real GCS connection (use RUN_GCS_SMOKE=1)",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
